@@ -18,40 +18,40 @@ publish_to: 'none' # Remove this line if you wish to publish to pub.dev
 version: 1.0.0+2021061701
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: '>=2.12.0 <3.0.0'
 
 dependencies:
   flutter:
     sdk: flutter
 
-  auto_route: ^0.5.0
-  flutter_udid: ^1.0.1
-  package_info: ^0.4.1
-  device_info: ^0.4.2+4
-  flutter_secure_storage: ^3.3.3
-  flutter_phoenix: ^0.1.0
-  webview_flutter: ^0.3.22+1
-  http: ^0.12.1
-  provider: ^4.1.3
-  flutter_spinkit: "^4.1.2"
-  shared_preferences: ^0.5.12+4
-  connectivity: ^0.4.2
-  badges: ^1.2.0
-  font_awesome_flutter: ^8.8.1
-  carousel_slider: ^2.2.1
-  url_launcher: ^5.5.2
-  scrollable_positioned_list: ^0.1.7
+  auto_route: ^2.2.0
+  flutter_udid: ^2.0.0
+  package_info: ^2.0.2
+  device_info: ^2.0.2
+  flutter_secure_storage: ^4.2.0
+  flutter_phoenix: ^1.0.0
+  webview_flutter: ^2.0.9
+  http: ^0.13.3
+  provider: ^5.0.0
+  flutter_spinkit: ^5.0.0
+  shared_preferences: ^2.0.6
+  connectivity: ^3.0.6
+  badges: ^2.0.1
+  font_awesome_flutter: ^9.1.0
+  carousel_slider: ^4.0.0
+  url_launcher: ^6.0.7
+  scrollable_positioned_list: ^0.2.0-nullsafety.0
 
   # The following adds the Cupertino Icons font to your application.
   # Use with the CupertinoIcons class for iOS style icons.
-  cupertino_icons: ^1.0.0
+  cupertino_icons: ^1.0.3
 
 dev_dependencies:
   flutter_test:
     sdk: flutter
 
-  build_runner:
-  auto_route_generator: ^0.5.0
+  build_runner: ^2.0.5
+  auto_route_generator: ^2.1.0
 
 # For information on the generic Dart part of this file, see the
 # following page: https://dart.dev/tools/pub/pubspec
