@@ -2,9 +2,9 @@ import 'package:flutter/material.dart';
 import 'package:sbit_mobile/Helper/AppTheme/appColors.dart';
 
 class DrawerTile extends StatelessWidget {
-  final String title;
-  final IconData icon;
-  final Function onTap;
+  final String? title;
+  final IconData? icon;
+  final Function? onTap;
   final bool isSelected;
   final Color iconColor;
 
@@ -18,7 +18,7 @@ class DrawerTile extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
     return GestureDetector(
-      onTap: onTap,
+      onTap: onTap as void Function()?,
       child: Container(
           padding: EdgeInsets.all(16),
           decoration: BoxDecoration(
@@ -30,7 +30,7 @@ class DrawerTile extends StatelessWidget {
                   size: 20, color: isSelected ? AppColors.white : iconColor),
               Padding(
                 padding: const EdgeInsets.only(left: 8.0),
-                child: Text(title,
+                child: Text(title!,
                     style: TextStyle(
                         letterSpacing: .3,
                         fontSize: 15,
