@@ -2,20 +2,20 @@ import 'package:flutter/material.dart';
 import 'package:sbit_mobile/Helper/AppTheme/appColors.dart';
 
 class TableCellSettings extends StatelessWidget {
-  final String title;
-  final Function onTap;
+  final String? title;
+  final Function? onTap;
   TableCellSettings({this.title, this.onTap});
 
   @override
   Widget build(BuildContext context) {
     return GestureDetector(
-      onTap: onTap,
+      onTap: onTap as void Function()?,
       child: Padding(
         padding: const EdgeInsets.only(top: 16.0),
         child: Row(
           mainAxisAlignment: MainAxisAlignment.spaceBetween,
           children: [
-            Text(title, style: TextStyle(color: AppColors.text)),
+            Text(title!, style: TextStyle(color: AppColors.text)),
             Padding(
               padding: const EdgeInsets.only(right: 8.0),
               child: Icon(Icons.arrow_forward_ios,
