@@ -3,13 +3,13 @@ import 'package:sbit_mobile/Helper/AppTheme/appColors.dart';
 import 'package:sbit_mobile/Helper/Component/input.dart';
 
 class InputDialog extends StatelessWidget {
-  final String titleMsg;
-  final String contentMsg;
-  final Function onRetryPressed1;
-  final Function onRetryPressed;
-  final Input textField;
+  final String? titleMsg;
+  final String? contentMsg;
+  final Function? onRetryPressed1;
+  final Function? onRetryPressed;
+  final Input? textField;
 
-  const InputDialog({Key key, this.titleMsg, this.contentMsg, this.onRetryPressed, this.onRetryPressed1, this.textField}) : super(key: key);
+  const InputDialog({Key? key, this.titleMsg, this.contentMsg, this.onRetryPressed, this.onRetryPressed1, this.textField}) : super(key: key);
   
   @override
   Widget build(BuildContext context) {
@@ -49,7 +49,7 @@ class InputDialog extends StatelessWidget {
               alignment: Alignment.centerLeft,
               width: double.infinity,
               child: Text(
-                titleMsg, 
+                titleMsg!, 
                 style: TextStyle(
                   fontWeight: FontWeight.bold, 
                   fontSize: 20.0,
@@ -64,7 +64,7 @@ class InputDialog extends StatelessWidget {
               child: FittedBox(
                 fit: BoxFit.fitWidth,
                 child: Text(
-                  contentMsg, 
+                  contentMsg!, 
                   style: TextStyle(
                     fontWeight: FontWeight.normal, 
                     fontSize: 18.0,
@@ -97,7 +97,7 @@ class InputDialog extends StatelessWidget {
                             color: Colors.pink,
                           ),
                         ),
-                      onPressed: onRetryPressed1,
+                      onPressed: onRetryPressed1 as void Function()?,
                     ),
                   ),
                   SizedBox(width: 8,),
@@ -109,7 +109,7 @@ class InputDialog extends StatelessWidget {
                       color: AppColors.label,
                       textColor: Colors.white,
                       shape: RoundedRectangleBorder(borderRadius: new BorderRadius.circular(6.0)),
-                      onPressed: onRetryPressed,
+                      onPressed: onRetryPressed as void Function()?,
                     ),
                   ),
                 ],
@@ -138,7 +138,7 @@ class InputDialog extends StatelessWidget {
             children: [
               Padding(
                 padding: const EdgeInsets.only(left: 16.0, right: 16.0, top: 24.0), 
-                child: Text(titleMsg, style: TextStyle(fontWeight: FontWeight.bold, fontSize: 24.0),),
+                child: Text(titleMsg!, style: TextStyle(fontWeight: FontWeight.bold, fontSize: 24.0),),
               ),
             ],
           ),
@@ -147,7 +147,7 @@ class InputDialog extends StatelessWidget {
             children: [
             Padding(
                 padding: const EdgeInsets.only(left: 16.0, right: 16.0, top: 10.0), 
-                child: Text(contentMsg, style: TextStyle(fontWeight: FontWeight.w500, fontSize: 19.0),),
+                child: Text(contentMsg!, style: TextStyle(fontWeight: FontWeight.w500, fontSize: 19.0),),
               ),
             ],
           ),
@@ -170,7 +170,7 @@ class InputDialog extends StatelessWidget {
                             color: Color(0xff44D62C),
                           ),
                         ),
-                      onPressed: onRetryPressed1,
+                      onPressed: onRetryPressed1 as void Function()?,
                     ),
                   ),
                   SizedBox(width: 8,),
@@ -182,7 +182,7 @@ class InputDialog extends StatelessWidget {
                       color: Color(0xff44D62C),
                       textColor: Colors.white,
                       shape: RoundedRectangleBorder(borderRadius: new BorderRadius.circular(3.0)),
-                      onPressed: onRetryPressed,
+                      onPressed: onRetryPressed as void Function()?,
                     ),
                   ),
                 ],
