@@ -3,15 +3,15 @@ import 'package:provider/provider.dart';
 import 'package:sbit_mobile/Helper/AppTheme/appColors.dart';
 import 'package:sbit_mobile/Helper/Provider/counter_bloc.dart';
 
-CounterBloc counterBloc;
+late CounterBloc counterBloc;
 
 class QuantityDialog extends StatelessWidget {
-  final String titleMsg;
-  final String contentMsg;
-  final Function onRetryPressed1;
-  final Function onRetryPressed;
+  final String? titleMsg;
+  final String? contentMsg;
+  final Function? onRetryPressed1;
+  final Function? onRetryPressed;
 
-  const QuantityDialog({Key key, this.titleMsg, this.contentMsg, this.onRetryPressed, this.onRetryPressed1}) : super(key: key);
+  const QuantityDialog({Key? key, this.titleMsg, this.contentMsg, this.onRetryPressed, this.onRetryPressed1}) : super(key: key);
   
   @override
   Widget build(BuildContext context) {
@@ -52,7 +52,7 @@ class QuantityDialog extends StatelessWidget {
               alignment: Alignment.centerLeft,
               width: double.infinity,
               child: Text(
-                titleMsg, 
+                titleMsg!, 
                 style: TextStyle(
                   fontWeight: FontWeight.bold, 
                   fontSize: 20.0,
@@ -65,7 +65,7 @@ class QuantityDialog extends StatelessWidget {
               alignment: Alignment.topLeft,
               width: double.infinity,
               child: Text(
-                contentMsg, 
+                contentMsg!, 
                 style: TextStyle(
                   fontWeight: FontWeight.normal, 
                   fontSize: 18.0,
@@ -114,7 +114,7 @@ class QuantityDialog extends StatelessWidget {
                             color: Colors.pink,
                           ),
                         ),
-                      onPressed: onRetryPressed1,
+                      onPressed: onRetryPressed1 as void Function()?,
                     ),
                   ),
                   SizedBox(width: 8,),
@@ -126,7 +126,7 @@ class QuantityDialog extends StatelessWidget {
                       color: AppColors.label,
                       textColor: Colors.white,
                       shape: RoundedRectangleBorder(borderRadius: new BorderRadius.circular(6.0)),
-                      onPressed: onRetryPressed,
+                      onPressed: onRetryPressed as void Function()?,
                     ),
                   ),
                 ],
@@ -155,7 +155,7 @@ class QuantityDialog extends StatelessWidget {
             children: [
               Padding(
                 padding: const EdgeInsets.only(left: 16.0, right: 16.0, top: 24.0), 
-                child: Text(titleMsg, style: TextStyle(fontWeight: FontWeight.bold, fontSize: 24.0),),
+                child: Text(titleMsg!, style: TextStyle(fontWeight: FontWeight.bold, fontSize: 24.0),),
               ),
             ],
           ),
@@ -164,7 +164,7 @@ class QuantityDialog extends StatelessWidget {
             children: [
             Padding(
                 padding: const EdgeInsets.only(left: 16.0, right: 16.0, top: 10.0), 
-                child: Text(contentMsg, style: TextStyle(fontWeight: FontWeight.w500, fontSize: 19.0),),
+                child: Text(contentMsg!, style: TextStyle(fontWeight: FontWeight.w500, fontSize: 19.0),),
               ),
             ],
           ),
@@ -187,7 +187,7 @@ class QuantityDialog extends StatelessWidget {
                             color: Color(0xff44D62C),
                           ),
                         ),
-                      onPressed: onRetryPressed1,
+                      onPressed: onRetryPressed1 as void Function()?,
                     ),
                   ),
                   SizedBox(width: 8,),
@@ -199,7 +199,7 @@ class QuantityDialog extends StatelessWidget {
                       color: Color(0xff44D62C),
                       textColor: Colors.white,
                       shape: RoundedRectangleBorder(borderRadius: new BorderRadius.circular(3.0)),
-                      onPressed: onRetryPressed,
+                      onPressed: onRetryPressed as void Function()?,
                     ),
                   ),
                 ],
