@@ -23,7 +23,7 @@ class CardSquare extends StatelessWidget {
         height: 250,
         width: null,
         child: GestureDetector(
-          onTap: tap,
+          onTap: tap as void Function()?,
           child: Card(
               elevation: 0.4,
               shape: RoundedRectangleBorder(
