@@ -8,8 +8,8 @@ import 'package:sbit_mobile/Helper/ShowDialog/quantity_dialog.dart';
 
 class DialogHelper {
   static loadingDialog(context) => showDialog(context:context, barrierDismissible: false, builder: (context) => LoadingDialog());
-  static customDialog(context, String title, String content, Function onPressed) => showDialog(context: context, barrierDismissible: false, builder: (context) => CustomDialog(titleMsg: title, contentMsg: content, onRetryPressed: onPressed));
+  static customDialog(context, String? title, String? content, Function onPressed) => showDialog(context: context, barrierDismissible: false, builder: (context) => CustomDialog(titleMsg: title, contentMsg: content, onRetryPressed: onPressed));
   static inputDialog(context, String title, String content, Function onPressed1, Function onPressed, Input costField) => showDialog(context: context, barrierDismissible: false, builder: (context) => InputDialog(titleMsg: title, contentMsg: content, onRetryPressed1: onPressed1, onRetryPressed: onPressed, textField: costField,));
-  static quantityDialog(context, String title, String content, Function onPressed1, Function onPressed) => showDialog(context: context, barrierDismissible: false, builder: (context) => QuantityDialog(titleMsg: title, contentMsg: content, onRetryPressed1: onPressed1, onRetryPressed: onPressed,));
+  static quantityDialog(context, String? title, String content, Function onPressed1, Function onPressed) => showDialog(context: context, barrierDismissible: false, builder: (context) => QuantityDialog(titleMsg: title, contentMsg: content, onRetryPressed1: onPressed1, onRetryPressed: onPressed,));
   static confirmDialog(context, String title, String content, Function onPressed1, Function onPressed) => showDialog(context: context, barrierDismissible: false, builder: (context) => ConfirmDialog(titleMsg: title, contentMsg: content, onRetryPressed1: onPressed1, onRetryPressed: onPressed,));
 }
